+++ Samples~/Scripts/SaintsEditor/Issues/Issue52.cs
@@ -1,3 +1,4 @@
+using System;
 using SaintsField.Playa;
 
 namespace SaintsField.Samples.Scripts.SaintsEditor.Issues
@@ -19,5 +20,27 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Issues
         [Layout("Layout")] public string layout3;
 
         public string end;
+
+        [Serializable]
+        public struct MyStruct
+        {
+            public string start;
+
+            [LayoutGroup("Group", ELayout.FoldoutBox, marginTop: 10, marginBottom: 10)]
+            public string group1;
+            public string group2;
+            public string group3;
+            [LayoutEnd("Group")]
+
+            public string middle;
+
+            [Layout("Layout", ELayout.TitleBox, marginTop: 10, marginBottom: 10)] public string layout1;
+            [Layout("Layout")] public string layout2;
+            [Layout("Layout")] public string layout3;
+
+            public string end;
+        }
+
+        [SaintsRow] public MyStruct myStruct;
     }
 }